backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6
langchain-core==0.1.10
langgraph==0.0.20
transformers==4.36.2
torch==2.1.2
pymupdf==1.23.8
google-generativeai==0.3.2
chromadb==0.4.22
